+++ src/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TransformStamped.java
@@ -14,7 +14,7 @@ import edu.wpi.rail.jrosbridge.messages.std.Header;
  * information.
  * 
  * @author Russell Toris -- rctoris@wpi.edu
- * @version March 6, 2014
+ * @version March 8, 2014
  */
 public class TransformStamped extends Message {
 
@@ -104,12 +104,12 @@ public class TransformStamped extends Message {
 	}
 
 	/**
-	 * Create a deep clone of this TransformStamped.
+	 * Create a clone of this TransformStamped.
 	 */
 	@Override
 	public TransformStamped clone() {
-		return new TransformStamped(this.header.clone(), new String(
-				this.childFrameID), this.transform.clone());
+		return new TransformStamped(this.header, this.childFrameID,
+				this.transform);
 	}
 
 	/**