+++ test/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TestQuaternionStamped.java
@@ -108,8 +108,6 @@ public class TestQuaternionStamped {
 		assertNotSame(p1, clone);
 		assertNotSame(p1.toString(), clone.toString());
 		assertNotSame(p1.toJsonObject(), clone.toJsonObject());
-		assertNotSame(p1.getHeader(), clone.getHeader());
-		assertNotSame(p1.getQuaternion(), clone.getQuaternion());
 	}
 
 	@Test